requirements.txt
mysql-connector==2.1.6
numpy==1.16.2
pandas==0.24.1
patsy==0.5.1
python-dateutil==2.8.0
pytz==2018.9
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
sklearn==0.0
statsmodels==0.9.0
